Implement to join battle for friend Fixes #209
